+++ analog/__init__.py
@@ -1,6 +1,16 @@
 """Analog - Log Analysis Utitliy."""
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
+import logging
+
+
+__version__ = '0.3.2.dev'
+
+
+# analog logger
+LOG = logging.getLogger('analog')
+LOG.addHandler(logging.NullHandler)
+
 
 from analog.analyzer import Analyzer, analyze
 from analog.exceptions import (AnalogError, InvalidFormatExpressionError,
@@ -9,9 +19,6 @@ from analog.formats import LogFormat
 from analog.main import main
 from analog.report import Report
 from analog.renderers import Renderer
-
-
-__version__ = '0.3.2.dev'
 
 
 __all__ = (